fix duplicate auth code
