+++ frontend/src/app/(admin)/admin-portal/page.tsx
@@ -1,30 +1,8 @@
 "use client";
-import { onAuthStateChanged } from "firebase/auth";
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-
 import { AdminSidebar } from "@/components/admin-portal/AdminSidebar";
 import { Section } from "@/components/admin-portal/Section";
-import { auth } from "@/firebase/firebase";
 
 export default function AdminPortal() {
-  const router = useRouter();
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      if (!user) {
-        router.push("/login");
-      } else {
-        setLoading(false);
-      }
-    });
-
-    return () => unsubscribe();
-  }, [router]);
-
-  if (loading) return null;
-
   const HomeComponents = [
     {
       name: "Rotating Cities",
