+++ frontend/src/app/(admin)/layout.tsx
@@ -2,6 +2,8 @@ import { Geist, Geist_Mono } from "next/font/google";
 
 import type { Metadata } from "next";
 
+import { ProtectedLayout } from "@/components/admin-portal/ProtectedLayout";
+
 import "./globals.css";
 
 const geistSans = Geist({
@@ -29,7 +31,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" className="scroll-smooth">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>{children}</body>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+        <ProtectedLayout>{children}</ProtectedLayout>
+      </body>
     </html>
   );
 }
