+++ frontend/src/components/admin-portal/ProtectedLayout.tsx
@@ -1,0 +1,28 @@
+"use client";
+
+import { onAuthStateChanged } from "firebase/auth";
+import { useRouter } from "next/navigation";
+import { useEffect, useState } from "react";
+
+import { auth } from "@/firebase/firebase";
+
+export function ProtectedLayout({ children }: { children: React.ReactNode }) {
+  const router = useRouter();
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      if (!user) {
+        router.push("/login");
+      } else {
+        setLoading(false);
+      }
+    });
+
+    return () => unsubscribe();
+  }, [router]);
+
+  if (loading) return null;
+
+  return <>{children}</>;
+}
